Allow redirect_uri to be provided at request time redirect_uri becomes an optional parameter to the SSOBase constructor, defaulting to None. redirect_uri is added as an optional parameter to get_login_url() and get_login_redirect(); if provided, it will override any value provided at construct time.
This allows code to work properly when it may run in multiple host environments (e.g., http://localhost:8000
and https://mydomain.com), or when it is running behind a reverse proxy like nginx and the
base url is not known until request time.

The API changes are backward-compatible so existing users should not be affected.

The examples and documentation are updated to show the new calling convention, which
is insensitive to host base URL.

Also includes a couple of very minor edits to suppress two pylint warnings that were present
before this commit.