Use dependency injection to consume SSO
The SSO object keeps oauth_client, which keeps the token during
verify_and_process. If there are 2 incoming callback requests at
roughtly the same time, it is possible that the token gets overwritten
by the second request so the user data fetch of the first request won't
succeed (or it will, returning the second user's data).